--- base_decorator.rb
@@ -1,14 +1,0 @@
-require_relative 'nameable'
-
-class Decorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable
-  end
-end
