--- capitalize_decorator.rb
@@ -1,7 +1,0 @@
-require_relative 'base_decorator'
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
